+++ packages/ws-worker/test/util/versions.test.ts
@@ -5,10 +5,10 @@ import calculateVersionString from '../../src/util/versions';
 // keys in this obejct are scrambled on purpose
 const versions = {
   worker: '2',
-  compiler: '5',
+  // compiler: '5',
   node: '1',
   engine: '3',
-  runtime: '4',
+  // runtime: '4',
 };
 
 // Util function to parse a version string into something easier to test
@@ -35,9 +35,7 @@ test('calculate version string', (t) => {
     `Versions for run run-1:
     ▸ node.js     1
     ▸ worker      2
-    ▸ engine      3
-    ▸ runtime     4
-    ▸ compiler    5`
+    ▸ engine      3`
   );
 });
 
@@ -49,8 +47,6 @@ test('helper should parse a version string and return the correct order', (t) =>
     ['node.js', '1'],
     ['worker', '2'],
     ['engine', '3'],
-    ['runtime', '4'],
-    ['compiler', '5'],
   ]);
 });
 
@@ -63,8 +59,6 @@ test("show unknown if a version isn't passed", (t) => {
     ['node.js', 'unknown'],
     ['worker', 'unknown'],
     ['engine', 'unknown'],
-    ['runtime', 'unknown'],
-    ['compiler', 'unknown'],
   ]);
 });
 
@@ -75,7 +69,7 @@ test('show adaptors last', (t) => {
   });
 
   const parsed = parse(str);
-  const common = parsed[5];
+  const common = parsed[3];
   t.deepEqual(common, ['@openfn/language-common', '1.0.0']);
 });
 
@@ -89,9 +83,9 @@ test('sort and list multiple adaptors', (t) => {
 
   const parsed = parse(str);
 
-  const a = parsed[5];
-  const j = parsed[6];
-  const z = parsed[7];
+  const a = parsed[3];
+  const j = parsed[4];
+  const z = parsed[5];
 
   t.deepEqual(a, ['a', '1']);
   t.deepEqual(j, ['j', '2']);