+++ packages/ws-worker/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ws-worker
 
+## 0.3.1
+
+### Patch Changes
+
+- Don't log compiler and runtime version logs
+
 ## 0.3.0
 
 ### Minor Changes